+++ lib/handlers/processResume.ts
@@ -1,35 +1,12 @@
-import {
-  createProcessActionAddDeletedEntityHandler,
-  createProcessActionHandler,
-} from './processActionHandler';
-import {
-  PROCESS_RESUME_ON,
-  PROCESS_RESUME_CASCADE,
-  PROCESS_RESUME_IF,
-  LOG_MESSAGES,
-} from '../constants';
-
-const action = 'resume';
+import { createProcessActionHandler } from './processActionHandler';
 
 export const handleProcessResume = createProcessActionHandler({
-  action: action,
-  annotations: {
-    ON: PROCESS_RESUME_ON,
-    CASCADE: PROCESS_RESUME_CASCADE,
-    IF: PROCESS_RESUME_IF,
-  },
+  action: 'resume',
   logMessages: {
-    NOT_TRIGGERED: LOG_MESSAGES.PROCESS_NOT_RESUMED,
+    NOT_TRIGGERED: 'Not resuming process as resume condition(s) are not met.',
     FETCH_FAILED: 'Failed to fetch entity for process resume.',
     INVALID_KEY: 'Failed to build business key for process resume.',
     EMPTY_KEY: 'Business key is empty for process resume.',
     FAILED: 'Failed to resume process with business key',
-  },
-});
-
-export const addDeletedEntityToRequestResume = createProcessActionAddDeletedEntityHandler({
-  action: action,
-  annotations: {
-    IF: PROCESS_RESUME_IF,
   },
 });