+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 - Reducing the context sent to the workflow API to only needed fields
 - Allowing the import of downloaded process models
+- Support for multiple start annotations on the same entity using qualifier
+- Support for multiple resume, cancel, and suspend annotations on the same entity using qualifier
 
 ## Version 0.1.1 - 2026-03-27